Restructured rosflight package include structure
