+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "rosflight/include/mavlink/v1.0"]
-	path = rosflight/include/mavlink/v1.0
+	path = rosflight/include/rosflight/mavlink/v1.0
 	url = https://github.com/rosflight/mavlink_c_library.git
