+++ rosflight/include/rosflight/mavrosflight/mavlink_bridge.h
@@ -8,6 +8,6 @@
 #ifndef MAVROSFLIGHT_MAVLINK_BRIDGE_H
 #define MAVROSFLIGHT_MAVLINK_BRIDGE_H
 
-#include <mavlink/v1.0/rosflight/mavlink.h>
+#include <rosflight/mavlink/v1.0/rosflight/mavlink.h>
 
 #endif // MAVROSFLIGHT_MAVLINK_BRIDGE_H
