+++ rosflight/include/rosflight/mavrosflight/param_manager.h
@@ -6,11 +6,11 @@
 #ifndef MAVROSFLIGHT_PARAM_MANAGER_H
 #define MAVROSFLIGHT_PARAM_MANAGER_H
 
-#include <mavrosflight/mavlink_bridge.h>
-#include <mavrosflight/mavlink_listener_interface.h>
-#include <mavrosflight/mavlink_serial.h>
-#include <mavrosflight/param.h>
-#include <mavrosflight/param_listener_interface.h>
+#include <rosflight/mavrosflight/mavlink_bridge.h>
+#include <rosflight/mavrosflight/mavlink_listener_interface.h>
+#include <rosflight/mavrosflight/mavlink_serial.h>
+#include <rosflight/mavrosflight/param.h>
+#include <rosflight/mavrosflight/param_listener_interface.h>
 
 #include <map>
 #include <string>
