+++ rosflight/include/rosflight/mavrosflight/sensors/mag.h
@@ -6,7 +6,7 @@
 #ifndef MAVROSFLIGHT_SENSORS_MAG_H
 #define MAVROSFLIGHT_SENSORS_MAG_H
 
-#include <mavrosflight/mavlink_bridge.h>
+#include <rosflight/mavrosflight/mavlink_bridge.h>
 #include <math.h>
 #include <eigen3/Eigen/Eigen>
 #include <random>
@@ -103,14 +103,14 @@ private:
 
   /*
       This function gets ellipsoid parameters via least squares on ellipsoidal data
-      according to the paper: Li, Qingde, and John G. Griffiths. "Least squares ellipsoid 
+      according to the paper: Li, Qingde, and John G. Griffiths. "Least squares ellipsoid
       specific fitting." Geometric modeling and processing, 2004. proceedings. IEEE, 2004.
   */
   Eigen::MatrixXd ellipsoidLS(std::deque<Eigen::Vector3d> meas);
 
   /*
       This function compute magnetometer calibration parameters according to Section 5.3 of the
-      paper: Renaudin, Valérie, Muhammad Haris Afzal, and Gérard Lachapelle. "Complete triaxis 
+      paper: Renaudin, Valérie, Muhammad Haris Afzal, and Gérard Lachapelle. "Complete triaxis
       magnetometer calibration in the magnetic domain." Journal of sensors 2010 (2010).
   */
   void magCal(Eigen::MatrixXd u, Eigen::MatrixXd &A, Eigen::MatrixXd &bb);
