+++ rosflight/include/rosflight/mavrosflight/serial_exception.h
